General lint and code cleaning
