+++ ring/basic_lifecycler_delegates_test.go
@@ -8,12 +8,13 @@ import (
 	"time"
 
 	"github.com/go-kit/kit/log"
-	"github.com/grafana/dskit/concurrency"
-	"github.com/grafana/dskit/services"
-	"github.com/grafana/dskit/test"
 	"github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
+	"github.com/grafana/dskit/concurrency"
+	"github.com/grafana/dskit/services"
+	"github.com/grafana/dskit/test"
 )
 
 func TestLeaveOnStoppingDelegate(t *testing.T) {
