+++ ring/client/pool_test.go
@@ -2,17 +2,18 @@ package client
 
 import (
 	"context"
-	fmt "fmt"
+	"fmt"
 	"testing"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/gogo/status"
-	"github.com/grafana/dskit/services"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/health/grpc_health_v1"
+
+	"github.com/grafana/dskit/services"
 )
 
 type mockClient struct {
