+++ ring/model.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/gogo/protobuf/proto"
+
 	"github.com/grafana/dskit/kv/codec"
 	"github.com/grafana/dskit/kv/memberlist"
 )
@@ -465,6 +466,7 @@ func (d *Desc) getTokensByZone() map[string][]uint32 {
 
 type CompareResult int
 
+// CompareResult responses
 const (
 	Equal                       CompareResult = iota // Both rings contain same exact instances.
 	EqualButStatesAndTimestamps                      // Both rings contain the same instances with the same data except states and timestamps (may differ).
