--- ring/util/http.go
@@ -1,39 +1,0 @@
-package util
-
-import (
-	"encoding/json"
-	"html/template"
-	"net/http"
-	"strings"
-)
-
-// RenderHTTPResponse either responds with json or a rendered html page using the passed in template
-// by checking the Accepts header
-func RenderHTTPResponse(w http.ResponseWriter, v interface{}, t *template.Template, r *http.Request) {
-	accept := r.Header.Get("Accept")
-	if strings.Contains(accept, "application/json") {
-		WriteJSONResponse(w, v)
-		return
-	}
-
-	err := t.Execute(w, v)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
-}
-
-// WriteJSONResponse writes some JSON as a HTTP response.
-func WriteJSONResponse(w http.ResponseWriter, v interface{}) {
-	w.Header().Set("Content-Type", "application/json")
-
-	data, err := json.Marshal(v)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	// We ignore errors here, because we cannot do anything about them.
-	// Write will trigger sending Status code, so we cannot send a different status code afterwards.
-	// Also this isn't internal error, but error communicating with client.
-	_, _ = w.Write(data)
-}
